pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prosaic-app"
version = "1.3.3"
description = "A writer-first terminal writing app"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    { name = "Deepansh Khurana" }
]
keywords = ["writing", "terminal", "markdown", "tui"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Editors",
]
dependencies = [
    "textual[syntax]>=0.50.0",
    "click>=8.1.0",
    "platformdirs>=4.0.0",
    "gitpython>=3.1.0",
    "pyspellchecker>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "textual-dev>=1.0.0",
]
grammar = [
    "language-tool-python>=2.8.0",
]

[project.scripts]
prosaic = "prosaic.__main__:main"

[project.urls]
Homepage = "https://prosaic.dimwit.me"
Repository = "https://github.com/dimwitlabs/prosaic"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["build*", "tests*"]

[tool.setuptools.package-data]
prosaic = ["themes/*.tcss", "LICENSE", "REFERENCE"]

[tool.pytest.ini_options]
testpaths = ["tests"]
